blog stuff

+++ blog/generator.py
@@ -1,5 +1,8 @@
 '''
-python generator.py <filename>
+python generator.py <filename> [d]
+if d is given, then we will do a dry run. 
+We'll create the html page, but not any sections or tags, and we won't modify
+the index.html page either. 
 '''
 
 import sys
@@ -19,6 +22,13 @@ except IndexError:
     print("ERROR: Need filename argument.")
     exit(1)
 
+try:
+    dryrun = sys.argv[2]
+except Exception as e:
+    dryrun = False
+else:
+    dryrun = True # accept any string to indicate dryrun
+
 print("Filename: " + filename)
 
 file = open(filename)
@@ -32,10 +42,37 @@ except Exception as e:
 
 print("Week: " + week)
 
-text = file.read().strip()
+lines = file.readlines()
 file.close()
 
-sections = re.findall("%% ?begin ?(\[[^\]]*][^%]*)%% ?end", text, re.DOTALL)
+# sections = re.findall("%% ?begin ?(\[[^\]]*][^%]*)%% ?end", lines, re.DOTALL)
+
+sections = []
+
+# find the %% begin line
+li = 0
+section = ""
+while li < len(lines):
+    while li < len(lines):
+        if re.search("%% ?begin", lines[li]) is not None:
+            section += lines[li]
+            li += 1
+            break
+        li += 1
+
+    # find the %% end line
+    while li < len(lines):
+        if re.search("%% ?end", lines[li]) is not None:
+            li += 1
+            break
+        section += lines[li]
+        li += 1
+
+    section = re.sub("%% ?begin", "", section)
+    section = re.sub("%% ?end", "", section)
+    sections.append(section)
+    section = ""
+
 full_content = ""
 all_tags = []
 
@@ -49,18 +86,20 @@ for si in range(len(sections)):
     if text != "": 
         full_content += text
 
-        # write contents into the section file
-        file = open(sec_filename, "w")
-        file.write(text)
-        file.close()
+        # write contents into the section file, if not dryrun
+        if not dryrun:
+            file = open(sec_filename, "w")
+            file.write(text)
+            file.close()
 
-        # create or append to tag files
+        # create or append to tag files, if not dryrun
         for tag in tags:
             if tag != '':
                 all_tags.append(tag)
-                file = open(TAGS + "/" + tag, "a")
-                file.write(sec_filename.replace(SECTIONS + "/", "") + "\n")
-                file.close()
+                if not dryrun:
+                    file = open(TAGS + "/" + tag, "a")
+                    file.write(sec_filename.replace(SECTIONS + "/", "") + "\n")
+                    file.close()
 
 # now write the full contents of the blog post in one file
 blog_filename = POSTS + "/w-{week}".format(week=week)
@@ -69,8 +108,12 @@ file = open(blog_filename + ".md", "w")
 file.write(full_content)
 file.close()
 
+# find all the main headings
+headings = re.findall("[^#]###[^#](.*)", full_content)
+headings = [s.strip() for s in headings]
+
 # now write it in HTML format
-os.system("markdown2 {0}.md > {0}.html".format(blog_filename))
+os.system("markdown2 {0}.md --extras header-ids -x fenced-code-blocks > {0}.html".format(blog_filename))
 
 # read the html generated
 file = open(blog_filename + ".html", "r")
@@ -86,22 +129,27 @@ file.close()
 content = template.replace("%%WEEK_NUM%%", "{}/{}/{}".format(week[0:2], week[2:4], week[4:6]))
 content = content.replace("%%TAGS%%", ", ".join(all_tags))
 content = content.replace("%%CONTENT%%", html_content)
+hds = ""
+for h in headings:
+    hds += "<a href=\"#{headingsm}\" class=\"w3-bar-item w3-button w3-hide-small w3-hover-pink w3-black\"> &nbsp;{heading}</a>".format(heading=h, headingsm=h.lower().replace(" ", "-"))
+content = content.replace("%%HEADINGS%%", hds)
 
 # write this content to the html file
 file = open(blog_filename + ".html", "w")
 file.write(content)
 file.close()
 
-# now add this blog post link to the home page (index.html)
-file = open(INDEX_HTML, "r")
-orig = file.read()
-file.close()
+# now add this blog post link to the home page (index.html), if not dryrun
+if not dryrun:
+    file = open(INDEX_HTML, "r")
+    orig = file.read()
+    file.close()
 
-new_blog = "{repl}\n<tr>\n<td><a href=\"blog/posts/w-{week}.html\">Week {y}/{m}/{d}</a></td>\n<td>{tags}</td>\n</tr>\n"
-new_blog = new_blog.format(week=week, y=week[0:2], m=week[2:4], d=week[4:6], tags=", ".join(all_tags), repl=BLOG_REPLACEMENT)
+    new_blog = "{repl}\n<tr>\n<td><a href=\"blog/posts/w-{week}.html\">Week {y}/{m}/{d}</a></td>\n<td>{tags}</td>\n</tr>\n"
+    new_blog = new_blog.format(week=week, y=week[0:2], m=week[2:4], d=week[4:6], tags=", ".join(all_tags), repl=BLOG_REPLACEMENT)
 
-new_index_html = orig.replace(BLOG_REPLACEMENT, new_blog)
+    new_index_html = orig.replace(BLOG_REPLACEMENT, new_blog)
 
-file = open(INDEX_HTML, "w")
-file.write(new_index_html)
-file.close()
+    file = open(INDEX_HTML, "w")
+    file.write(new_index_html)
+    file.close()
